pyproject.toml
[project]
name = "spotify-plus-mcp"
version = "0.2.0"
description = "MCP spotify project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "mcp==1.3.0",
 "python-dotenv>=1.0.1",
 "spotipy==2.24.0",
 "requests>=2.31.0",
]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.uv.sources]
spotify-plus-mcp = { workspace = true }

[project.scripts]
spotify-plus-mcp = "spotify_mcp:main"
spotify-plus-mcp-auth = "auth:main"
